+++ lib/basic_types.rb
@@ -1,0 +1,5 @@
+module LowType
+  class Boolean; end # TrueClass or FalseClass
+  class HTML < String; end
+  class JSON < String; end
+end