+++ lib/interfaces/adapter_interface.rb
@@ -1,0 +1,7 @@
+module LowType
+  class AdapterInterface
+    def redefine_methods
+      raise NotImplementedError
+    end
+  end
+end