+++ lib/local_types.rb
@@ -3,24 +3,24 @@ require_relative 'proxies/local_proxy'
 require_relative 'type_expression'
 require_relative 'value_expression'
 
-module TypeAssignment
+module LocalTypes
   class AssignmentError < StandardError; end
 
   def type(type_expression)
-    object = type_expression.default_value
+    referenced_object = type_expression.default_value
 
-    if !LowType.value?(object)
-      raise AssignmentError, "Single-instance objects like #{object} are not supported"
+    if !LowType.value?(referenced_object)
+      raise AssignmentError, "Single-instance objects like #{referenced_object} are not supported"
     end
 
     last_caller = caller_locations(1, 1).first
     file = LowType::FileProxy.new(path: last_caller.path, line: last_caller.lineno, scope: 'local type')
     local_proxy = LowType::LocalProxy.new(type_expression:, name: self, file:)
-    object.instance_variable_set('@local_proxy', local_proxy)
+    referenced_object.instance_variable_set('@local_proxy', local_proxy)
 
-    type_expression.validate!(value: object, proxy: local_proxy)
+    type_expression.validate!(value: referenced_object, proxy: local_proxy)
 
-    def object.with_type=(value)
+    def referenced_object.with_type=(value)
       local_proxy = self.instance_variable_get('@local_proxy')
       type_expression = local_proxy.type_expression
       type_expression.validate!(value:, proxy: local_proxy)
@@ -33,9 +33,9 @@ module TypeAssignment
       self
     end
 
-    return object.value if object.is_a?(ValueExpression)
+    return referenced_object.value if referenced_object.is_a?(ValueExpression)
 
-    object
+    referenced_object
   end
   alias_method :low_type, :type
 
@@ -43,19 +43,22 @@ module TypeAssignment
     LowType.value(type:)
   end
   alias_method :low_value, :value
-end
 
-module LowType
+  # Scoped to the class that includes LowTypes module.
   class Array < ::Array
-    def self.[](type)
-      return TypeExpression.new(type: [type]) if LowType.type?(type)
+    def self.[](*types)
+      if types.all? { |type| LowType.type?(type) }
+        return LowType::TypeExpression.new(type: [*types])
+      end
+
       super
     end
   end
 
+  # Scoped to the class that includes LowTypes module.
   class Hash < ::Hash
     def self.[](type)
-      return TypeExpression.new(type:) if LowType.type?(type)
+      return LowType::TypeExpression.new(type:) if LowType.type?(type)
       super
     end
   end